refactor: split Windows overlay lifecycle
